requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
rapidfuzz>=3.0.0
